switch to legacy_tox_ini, remove tox-ini-fmt.
I noticed https://github.com/tox-dev/tox-ini-fmt/issues/42, which makes
it hard to use comments in tox.ini. Auto-formatting tox.ini was the main
motivation to use a separate tox.ini at all, so just keep everything in
pyproject.toml.

Maybe tox will eventually have first-party support for pyproject.toml;
this will make the transition that much simpler.

It appears that tox==3.21.4 (appears on on ubuntu-22.04) can't actually
parse the legacy_tox_ini option, and so does not get to the point of
provisioning a tox environment for the approrpiate version. We now do
proper provisioning ourselves in github ci.