build(pyproject.toml): drop support for python 3.8
python 3.8 reached EOL on 2024-10-07. dropping support is a breaking change,
and we'll introduce other breaking changes shortly, so do this now.

this breaks use of the ubuntu-20.04 github actions os. that os will be
deprecated on 2025-04, so this is fine.

otherwise, the main changes are to typing. we also remove
backports.zoneinfo, which removes our only sys.version_info condition.

BREAKING CHANGE: drop support for python 3.8