refactor(resolver): make resolver generic
this makes resolver generic over its argument types, and not depend
directly on SubvolumeInfo/BackupInfo.

the goal is to support the planner refactor, allowing both
implementations to exist at the same time.

this requires any call to resolve() to supply a function to convert a
snapshot object to a backup object. I experimented with making resolve()
receive and/or return data in hermetic tuples, but this requires more
conversion code for all callers than this option.