tests: remove btrfsutil from some tests
this prepares for the upcoming removal of btrfsutil. I should have
removed btrfsutil from the resolver tests in the generic resolver
change.

the upcoming update2 tests will be created by copying the update tests,
and btrfsutil will be removed shortly after, so we want to remove
btrfsutil from these tests now to keep the changes as clean as possible.