refactor(planner): delete the old assessor logic
this deletes the old assessor logic, including the associated thunk utility and
the original "update" command. "btrfs2s3 update" now uses the new update2
command.

BREAKING CHANGE: if a source is newer than its snapshots, we unconditionally create a snapshot without user confirmation, even in an interactive terminal. new and old snapshots will be evaluated together with backups, and the user will be prompted for action. the user will have an option to undo the newly-created snapshots.